# openapi.yaml
openapi: 3.0.3
info:
  title: HR Agent AI — Backend API
  version: "1.1.0"
  description: |
    Backend для AI-ассистента «Техна» по HR-вопросам компании 1221 Системс.
    Полностью локальная архитектура: Node.js/Express + PostgreSQL + Ollama
    (qwen2.5:32b для генерации, bge-m3 для эмбеддингов) + локальный
    RAG-индекс по ЛНА (ПВТР, положения, инструкции).

    ## Аутентификация и роли

    Все защищённые эндпоинты принимают JWT в заголовке `Authorization: Bearer <token>`.
    JWT выдаётся `POST /api/auth/login` и содержит поле `role`:
      * `employee` — обычный сотрудник;
      * `admin` — администратор.

    Тестовые учётки:
      * `admin / admin` — администратор;
      * `12345 / password123` (Потапов), `67890 / password123` (Иванова),
        `11111 / password123` (Петров) — обычные сотрудники.

    ## Матрица доступа

    | Эндпоинт                              | Аноним | Сотрудник                | Админ |
    | ------------------------------------- | ------ | ------------------------ | ----- |
    | `GET  /health`                        |  ✅    |  ✅                      |  ✅    |
    | `POST /api/auth/register`             |  ✅    |  ✅                      |  ✅    |
    | `POST /api/auth/login`                |  ✅    |  ✅                      |  ✅    |
    | `POST /api/auth/verify`               |  ❌    |  ✅                      |  ✅    |
    | `POST /api/employee/auth`             |  ✅    |  ✅                      |  ✅    |
...

    Возможности:
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/employee/{id}:
    get:
      tags: [Employee]
      summary: Карточка сотрудника по табельному номеру
      description: |
        **Требует JWT.**
        * Сотрудник: разрешён только свой `id` (иначе 403).
        * Админ: любой `id`.
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeePublic"
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/employee/{id}/vacation:
    get:
      tags: [Employee]
      summary: Остаток отпуска и график
      description: |
        **Требует JWT.**
        * Сотрудник: только свой `id`.
        * Админ: любой `id`.
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VacationInfo"
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /api/employee/{id}/birthday:
    get:
      tags: [Employee]
      summary: Дата рождения и возраст
      description: |
        **Требует JWT.**
        * Сотрудник: только свой `id`.
        * Админ: любой `id`.
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  birthDate: { type: string, format: date, nullable: true }
                  age: { type: integer, nullable: true }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /api/employee/auth:
    post:
      tags: [Employee]
      summary: Лёгкая идентификация по табельному/e-mail (без пароля)
      description: |
        Публичный, без JWT. Возвращает opaque-токен (не JWT) для UX-сценария
        «вспомнить пользователя на этом устройстве». Никаких прав не выдаёт —
        для боевых вызовов используйте `POST /api/auth/login`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId: { type: string, example: "12345" }
                email: { type: string, example: "a.potapov@company.ru" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  employee: { $ref: "#/components/schemas/EmployeePublic" }
        "404": { $ref: "#/components/responses/Error" }
  /api/employee/search/by-name:
    get:
      tags: [Employee]
      summary: Поиск сотрудника по ФИО
      description: |
        **Требует JWT** (любая роль). Используется для вопросов вида
        «когда день рождения <имя коллеги>». Не отдаёт паролей/zarplata-полей.
      parameters:
        - in: query
          name: name
          required: true
          schema: { type: string }
          example: "Иванова"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items: { $ref: "#/components/schemas/EmployeePublic" }
        "401": { $ref: "#/components/responses/Error" }
...
  /api/knowledge:
    get:
      tags: [Knowledge]
      summary: Базовая «быстрая» KB + статистика RAG-индекса
      description: "**Требует JWT** (любая роль)."
      responses:
        "200":
          description: OK
        "401": { $ref: "#/components/responses/Error" }
  /api/knowledge/search:
    post:
      tags: [Knowledge]
      summary: Поиск по базе знаний (вектор + keyword fallback)
      description: "**Требует JWT** (любая роль)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:    { type: string, example: "перенос отпуска" }
                topK:     { type: integer, default: 4 }
                minScore: { type: number,  default: 0.35 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      title:   { type: string }
                      content: { type: string }
                      source:  { type: string }
                  hits:
                    type: array
                    items: { $ref: "#/components/schemas/RagHit" }
        "401": { $ref: "#/components/responses/Error" }
  /api/knowledge/reindex:
    post:
      tags: [Knowledge]
      summary: Принудительная переиндексация папки RAG_DOCS_DIR — admin only
      description: "**Только admin.** Обычный сотрудник получит 403."
      responses:
        "200":
          description: Готово
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:  { type: string }
                  indexed:  { type: integer }
                  chunks:   { type: integer }
                  total:    { type: integer }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /api/knowledge/index:
    get:
      tags: [Knowledge]
      summary: Состояние векторного индекса
      description: "**Требует JWT** (любая роль)."
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats:
                    type: object
                    properties:
                      total: { type: integer }
                      byDocument:
                        type: object
                        additionalProperties: { type: integer }
                  documents:
                    type: array
                    items:
                      type: object
                      properties:
                        id:         { type: string }
                        title:      { type: string }
                        type:       { type: string }
                        category:   { type: string }
                        sourcePath: { type: string }
                        chunkCount: { type: integer }
        "401": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EmployeeId:
      in: path
      name: id
      required: true
      schema: { type: string }
      example: "12345"
      description: |
        Табельный номер. Обычный сотрудник может указывать только собственный
        `employeeId` (иначе 403). Админ — любой существующий.
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:   { type: string }
        message: { type: string }
...
    EmployeePublic:
      type: object
      properties:
        id:           { type: string, example: "12345" }
        fullName:     { type: string, example: "Потапов Артем Павлович" }
        position:     { type: string }
        department:   { type: string }
        email:        { type: string }
        phone:        { type: string }
        hireDate:     { type: string, format: date, nullable: true }
        birthDate:    { type: string, format: date, nullable: true }
        vacationDays: { type: integer }
        nextVacation: { type: string, format: date, nullable: true }
        salary:       { type: integer, nullable: true }
    VacationInfo:
      type: object
      properties:
        remainingDays: { type: integer }
        nextVacation:  { type: string, format: date, nullable: true }
        vacationSchedule:
          type: array
          items:
            type: object
            properties:
              startDate: { type: string, format: date }
              endDate:   { type: string, format: date, nullable: true }
              days:      { type: integer }
              status:    { type: string, example: "planned" }
...
    RagHit:
      type: object
      properties:
        score:    { type: number, example: 0.66 }
        doc:      { type: string, example: "Правила внутреннего трудового распорядка" }
        section:  { type: string, example: "п. 7.6" }
...